Create libktx project (#1071)
Move the libktx build from the root CMakeLists.txt to its own in the
`lib` directory. Although still within KTX-Software, the new libktx
project `CMakeLists.txt` can be used as either a sub-project of
KTX-Software or as a standalone project to build just libktx.

This has been done due to both direct requests, such as
https://github.com/KhronosGroup/KTX-Software/discussions/995 and
indirect requests such as complaints that the root CMakeLists is too
complicated. The effort has led to some useful cleanup and better
modularization.

The following FEATURE options have been moved to the libktx project and
consequently renamed with a `LIBKTX` prefix:
- LIBKTX_FEATURE_KTX1 
- LIBKTX_FEATURE_KTX2
- LIBKTX_FEATURE_VK_UPLOAD
- LIBKTX_FEATURE_GL_UPLOAD
- LIBKTX_FEATURE_ETC_UNPACK

These options have been duplicated in the libktx project so they can be
set when it is the top-level project:
- LIBKTX_EMBED_BITCODE
- LIBKTX_WERROR

When libktx is used as a sub-project, they are set from the similarly
named KTX-Software options.

Add a workflow to test configuring and building a standalone libktx.

Along the way, fixed a bug, that caused configurations in the linux.yml
workflow that were supposed to incorporate OpenCL to not do so. However
it is not possible to run the tests on those configurations as the
Actions runners report the cpu type as "generic" so Portable OpenCL
(POCL) can't figure out what to do.

Also along the way, update the ignore lists in all the workflow files to
include newer workflows files.